Item moving system complete
Items can be moved to the crafting panel. If the same item is in that panel, the amount of the item dragged is added to the amount in the slot. If the items are not the same, the item gets returned to its original position. The item is also returned to the original position if the slot is not valid (random space).